+++ docs/reference/compound-list.js
@@ -1,5 +1,5 @@
 (function() {
   const script = document.getElementById('compound-list');
   const varname = script.getAttribute('data-varname');
-  this[varname] = ["struct_pico_test_metadata.md","logger_8c.md","fixtures_8c.md","filter_8c.md","hooks_8c.md","traverse_8c.md","main_suite_8inc.md","_r_e_a_d_m_e_8md.md","group__public__interface.md","tags_8c.md","group__assertions.md","group__test__cases.md","dir_d28a4824dc47e487b107a5db32ef43c4.md","group__fixtures.md","picotest_8h.md","group__test__suites.md","dir_d44c64559bbebec7f509842c48db8b23.md","indexpage.md"];
+  this[varname] = ["fixtures_8c.md","filter_8c.md","struct_pico_test_metadata.md","hooks_8c.md","main_suite_8inc.md","logger_8c.md","tags_8c.md","traverse_8c.md","group__fixtures.md","group__test__cases.md","dir_d44c64559bbebec7f509842c48db8b23.md","_r_e_a_d_m_e_8md.md","group__assertions.md","group__test__suites.md","dir_d28a4824dc47e487b107a5db32ef43c4.md","group__public__interface.md","picotest_8h.md","indexpage.md"];
   })();