+++ include/picotest.h
@@ -84,6 +84,73 @@
 /*! \} End of Version */
 
 /*!
+ * \name Namespaces
+ * \{
+ */
+
+/**
+ * Activate namespacing for the current file (disabled by default).
+ *
+ * Defining #PICOTEST_NAMESPACE before including `picotest.h` will use the
+ * provided value as a prefix for all the symbols in the file to avoid name
+ * clashes. This is useful with large test suites to keep local test names
+ * short.
+ */
+#ifdef DOXYGEN
+#define PICOTEST_NAMESPACE
+#endif
+
+/**
+ * String delimiter between namespace and test names (defaults to `"."`).
+ */
+#ifndef PICOTEST_NAMESPACE_DELIMITER
+#define PICOTEST_NAMESPACE_DELIMITER "."
+#endif
+
+/*! \cond IGNORE */
+#ifdef PICOTEST_NAMESPACE
+#define _PICOTEST_NAMESPACED(name)                                             \
+    _PICOTEST_CONCATENATE(PICOTEST_NAMESPACE, _PICOTEST_CONCATENATE(_, name))
+#define _PICOTEST_DISPLAY_NAME(name)                                           \
+    _PICOTEST_STRINGIZE(PICOTEST_NAMESPACE)                                    \
+    PICOTEST_NAMESPACE_DELIMITER _PICOTEST_STRINGIZE(name)
+#else
+#define _PICOTEST_NAMESPACED(name) name
+#define _PICOTEST_DISPLAY_NAME(name) _PICOTEST_STRINGIZE(name)
+#endif
+/*! \endcond */
+
+/**
+ * Export a test to make it globally accessible.
+ *
+ * This creates a global wrapper around the namespaced test, allowing it to be
+ * referenced externally.
+ *
+ * @param _testName     Test name.
+ */
+#ifdef PICOTEST_NAMESPACE
+#define PICOTEST_EXPORT(_testName)                                             \
+    int _testName(const char *cond) {                                          \
+        return _PICOTEST_NAMESPACED(_testName)(cond);                          \
+    }                                                                          \
+    extern PicoTestMetadata _PICOTEST_METADATA(                                \
+        _PICOTEST_NAMESPACED(_testName));                                      \
+    static PicoTestMetadata *_testName##_exported_subtests[] = {               \
+        &_PICOTEST_METADATA(_PICOTEST_NAMESPACED(_testName)), NULL};           \
+    PicoTestMetadata _PICOTEST_METADATA(_testName) = {                         \
+        _PICOTEST_STRINGIZE(_testName),                                        \
+        __FILE__,                                                              \
+        __LINE__,                                                              \
+        _testName,                                                             \
+        1,                                                                     \
+        (const struct PicoTestMetadata **)_testName##_exported_subtests};
+#else
+#define PICOTEST_EXPORT(_testName)
+#endif
+
+/*! \} End of Namespaces */
+
+/*!
  * \name Test Functions
  * \{
  */
@@ -126,8 +193,8 @@ typedef struct PicoTestMetadata {
  * @see PICOTEST_METADATA
  */
 #define PICOTEST_EXTERN(_testName)                                             \
-    extern PicoTestProc _testName;                                             \
-    extern PicoTestMetadata _PICOTEST_METADATA(_testName);
+    extern PicoTestProc _PICOTEST_NAMESPACED(_testName);                       \
+    extern PicoTestMetadata _PICOTEST_METADATA(_PICOTEST_NAMESPACED(_testName));
 
 /**
  * Get test metadata.
@@ -139,17 +206,19 @@ typedef struct PicoTestMetadata {
  * @see PicoTestMetadata
  * @see PICOTEST_EXTERN
  */
-#define PICOTEST_METADATA(_testName) &_PICOTEST_METADATA(_testName)
+#define PICOTEST_METADATA(_testName)                                           \
+    &_PICOTEST_METADATA(_PICOTEST_NAMESPACED(_testName))
 
 /*! \cond IGNORE */
-#define _PICOTEST_METADATA(_testName) _testName##_metadata
+#define _PICOTEST_METADATA(_testName)                                          \
+    _PICOTEST_CONCATENATE(_testName, _metadata)
 #define _PICOTEST_TEST_DECLARE(_testName, _nbSubtests, _subtests)              \
-    PicoTestProc _testName;                                                    \
-    PicoTestMetadata _PICOTEST_METADATA(_testName) = {                         \
-        _PICOTEST_STRINGIZE(_testName),                                        \
+    PicoTestProc _PICOTEST_NAMESPACED(_testName);                              \
+    PicoTestMetadata _PICOTEST_METADATA(_PICOTEST_NAMESPACED(_testName)) = {   \
+        _PICOTEST_DISPLAY_NAME(_testName),                                     \
         __FILE__,                                                              \
         __LINE__,                                                              \
-        _testName,                                                             \
+        _PICOTEST_NAMESPACED(_testName),                                       \
         _nbSubtests,                                                           \
         (const struct PicoTestMetadata **)_subtests};
 /*! \endcond */
@@ -530,18 +599,20 @@ static void _picoTest_logFailure(const char *file, int line, const char *type,
 /*! \cond IGNORE */
 #define _PICOTEST_CASE_DECLARE(_testName)                                      \
     _PICOTEST_TEST_DECLARE(_testName, 0, NULL);                                \
-    static int _testName##_testCaseRunner(void);                               \
-    int _testName(const char *cond) {                                          \
+    static int _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),          \
+                                     _testCaseRunner)(void);                   \
+    int _PICOTEST_NAMESPACED(_testName)(const char *cond) {                    \
         int fail = 0;                                                          \
         PicoTestFilterResult filterResult =                                    \
             (cond == NULL)                                                     \
                 ? PICOTEST_FILTER_PASS                                         \
-                : PICOTEST_FILTER(_testName, _PICOTEST_STRINGIZE(_testName),   \
-                                  cond);                                       \
+                : PICOTEST_FILTER(_PICOTEST_NAMESPACED(_testName),             \
+                                  _PICOTEST_DISPLAY_NAME(_testName), cond);    \
         switch (filterResult) {                                                \
         case PICOTEST_FILTER_PASS:                                             \
         case PICOTEST_FILTER_PASS_PROPAGATE:                                   \
-            fail += _testName##_testCaseRunner();                              \
+            fail += _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),     \
+                                          _testCaseRunner)();                  \
             break;                                                             \
         default:;                                                              \
         }                                                                      \
@@ -549,19 +620,20 @@ static void _picoTest_logFailure(const char *file, int line, const char *type,
     }
 
 #define _PICOTEST_CASE_RUNNER_BEGIN(_testName)                                 \
-    static int _testName##_testCaseRunner() {                                  \
+    static int _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),          \
+                                     _testCaseRunner)() {                      \
         int abort;                                                             \
         jmp_buf failureEnv;                                                    \
         jmp_buf *oldEnv = _picoTest_failureEnv;                                \
         int fail, oldFail = _picoTest_fail;                                    \
         _picoTest_failureEnv = &failureEnv;                                    \
         _picoTest_fail = 0;                                                    \
-        PICOTEST_CASE_ENTER(_PICOTEST_STRINGIZE(_testName));                   \
+        PICOTEST_CASE_ENTER(_PICOTEST_DISPLAY_NAME(_testName));                \
         abort = setjmp(failureEnv);
 
 #define _PICOTEST_CASE_RUNNER_END(_testName)                                   \
     fail = _picoTest_fail;                                                     \
-    PICOTEST_CASE_LEAVE(_PICOTEST_STRINGIZE(_testName), fail);                 \
+    PICOTEST_CASE_LEAVE(_PICOTEST_DISPLAY_NAME(_testName), fail);              \
     _picoTest_failureEnv = oldEnv;                                             \
     _picoTest_fail = oldFail;                                                  \
     return fail;                                                               \
@@ -569,41 +641,53 @@ static void _picoTest_logFailure(const char *file, int line, const char *type,
 
 #define _PICOTEST_CASE_1(_testName)                                            \
     _PICOTEST_CASE_DECLARE(_testName)                                          \
-    static void _testName##_testCase(void);                                    \
+    static void _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),         \
+                                      _testCase)(void);                        \
     _PICOTEST_CASE_RUNNER_BEGIN(_testName)                                     \
     if (!abort) {                                                              \
-        _testName##_testCase();                                                \
+        _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName), _testCase)();   \
     }                                                                          \
     _PICOTEST_CASE_RUNNER_END(_testName)                                       \
-    static void _testName##_testCase(void)
+    static void _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),         \
+                                      _testCase)(void)
 
 #define _PICOTEST_CASE_2(_testName, _fixtureName)                              \
     _PICOTEST_CASE_DECLARE(_testName)                                          \
-    static void _testName##_testCase(void);                                    \
+    static void _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),         \
+                                      _testCase)(void);                        \
     _PICOTEST_CASE_RUNNER_BEGIN(_testName)                                     \
     if (!abort) {                                                              \
-        _PICOTEST_FIXTURE_CALL_SETUP(_fixtureName, _testName, NULL);           \
-        _testName##_testCase();                                                \
+        _PICOTEST_FIXTURE_CALL_SETUP(_fixtureName,                             \
+                                     _PICOTEST_NAMESPACED(_testName), NULL);   \
+        _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName), _testCase)();   \
     }                                                                          \
-    _PICOTEST_FIXTURE_CALL_TEARDOWN(_fixtureName, _testName, NULL,             \
-                                    _picoTest_fail);                           \
+    _PICOTEST_FIXTURE_CALL_TEARDOWN(                                           \
+        _fixtureName, _PICOTEST_NAMESPACED(_testName), NULL, _picoTest_fail);  \
     _PICOTEST_CASE_RUNNER_END(_testName)                                       \
-    static void _testName##_testCase()
+    static void _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),         \
+                                      _testCase)()
 
 #define _PICOTEST_CASE_3(_testName, _fixtureName, _context)                    \
     _PICOTEST_CASE_DECLARE(_testName)                                          \
-    static void _testName##_testCase(struct _fixtureName##_Context *);         \
+    static void _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),         \
+                                      _testCase)(                              \
+        struct _fixtureName##_Context *);                                      \
     _PICOTEST_CASE_RUNNER_BEGIN(_testName) {                                   \
         struct _fixtureName##_Context context;                                 \
         if (!abort) {                                                          \
-            _PICOTEST_FIXTURE_CALL_SETUP(_fixtureName, _testName, &context);   \
-            _testName##_testCase(&context);                                    \
+            _PICOTEST_FIXTURE_CALL_SETUP(                                      \
+                _fixtureName, _PICOTEST_NAMESPACED(_testName), &context);      \
+            _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),             \
+                                  _testCase)(&context);                        \
         }                                                                      \
-        _PICOTEST_FIXTURE_CALL_TEARDOWN(_fixtureName, _testName, &context,     \
-                                        _picoTest_fail);                       \
+        _PICOTEST_FIXTURE_CALL_TEARDOWN(_fixtureName,                          \
+                                        _PICOTEST_NAMESPACED(_testName),       \
+                                        &context, _picoTest_fail);             \
     }                                                                          \
     _PICOTEST_CASE_RUNNER_END(_testName)                                       \
-    static void _testName##_testCase(struct _fixtureName##_Context *_context)
+    static void _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_testName),         \
+                                      _testCase)(                              \
+        struct _fixtureName##_Context * _context)
 /*! \endcond */
 
 /*! \} End of Test Case Definitions */
@@ -1391,46 +1475,56 @@ typedef void(PicoTestFixtureAfterTeardownProc)(const char *fixtureName,
  */
 #define PICOTEST_SUITE(_suiteName, ...)                                        \
     _PICOTEST_FOR_EACH(PICOTEST_EXTERN, __VA_ARGS__)                           \
-    static PicoTestMetadata *_suiteName##_subtests[] = {_PICOTEST_FOR_EACH(    \
-        _PICOTEST_SUITE_ENUMERATE_SUBTEST, __VA_ARGS__) NULL};                 \
-    _PICOTEST_TEST_DECLARE(_suiteName, _PICOTEST_ARGCOUNT(__VA_ARGS__),        \
-                           _suiteName##_subtests);                             \
-    static int _suiteName##_testSuiteRunner(const char *cond) {                \
+    static PicoTestMetadata *_PICOTEST_CONCATENATE(                            \
+        _PICOTEST_NAMESPACED(_suiteName), _subtests)[] = {                     \
+        _PICOTEST_FOR_EACH(_PICOTEST_SUITE_ENUMERATE_SUBTEST, __VA_ARGS__)     \
+            NULL};                                                             \
+    _PICOTEST_TEST_DECLARE(                                                    \
+        _suiteName, _PICOTEST_ARGCOUNT(__VA_ARGS__),                           \
+        _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_suiteName), _subtests));   \
+    static int _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_suiteName),         \
+                                     _testSuiteRunner)(const char *cond) {     \
         const int nb = _PICOTEST_ARGCOUNT(__VA_ARGS__);                        \
-        PicoTestMetadata **subtest = _suiteName##_subtests;                    \
+        PicoTestMetadata **subtest = _PICOTEST_CONCATENATE(                    \
+            _PICOTEST_NAMESPACED(_suiteName), _subtests);                      \
         int fail = 0;                                                          \
-        PICOTEST_SUITE_ENTER(_PICOTEST_STRINGIZE(_suiteName), nb);             \
+        PICOTEST_SUITE_ENTER(_PICOTEST_DISPLAY_NAME(_suiteName), nb);          \
         for (; *subtest; subtest++) {                                          \
-            const int index = (int)(subtest - _suiteName##_subtests);          \
+            const int index =                                                  \
+                (int)(subtest -                                                \
+                      _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_suiteName),  \
+                                            _subtests));                       \
             int sfail = 0;                                                     \
-            PICOTEST_SUITE_BEFORE_SUBTEST(_PICOTEST_STRINGIZE(_suiteName), nb, \
-                                          fail, index, (*subtest)->name);      \
+            PICOTEST_SUITE_BEFORE_SUBTEST(_PICOTEST_DISPLAY_NAME(_suiteName),  \
+                                          nb, fail, index, (*subtest)->name);  \
             sfail = (*subtest)->test(cond);                                    \
             fail += sfail;                                                     \
-            PICOTEST_SUITE_AFTER_SUBTEST(_PICOTEST_STRINGIZE(_suiteName), nb,  \
-                                         fail, index, (*subtest)->name,        \
+            PICOTEST_SUITE_AFTER_SUBTEST(_PICOTEST_DISPLAY_NAME(_suiteName),   \
+                                         nb, fail, index, (*subtest)->name,    \
                                          sfail);                               \
         }                                                                      \
-        PICOTEST_SUITE_LEAVE(_PICOTEST_STRINGIZE(_suiteName), nb, fail);       \
+        PICOTEST_SUITE_LEAVE(_PICOTEST_DISPLAY_NAME(_suiteName), nb, fail);    \
         return fail;                                                           \
     }                                                                          \
-    int _suiteName(const char *cond) {                                         \
+    int _PICOTEST_NAMESPACED(_suiteName)(const char *cond) {                   \
         int fail = 0;                                                          \
         PicoTestFilterResult filterResult =                                    \
             (cond == NULL)                                                     \
                 ? PICOTEST_FILTER_PASS                                         \
-                : PICOTEST_FILTER(_suiteName, _PICOTEST_STRINGIZE(_suiteName), \
-                                  cond);                                       \
+                : PICOTEST_FILTER(_PICOTEST_NAMESPACED(_suiteName),            \
+                                  _PICOTEST_DISPLAY_NAME(_suiteName), cond);   \
         switch (filterResult) {                                                \
         case PICOTEST_FILTER_PASS:                                             \
             cond = NULL;                                                       \
         case PICOTEST_FILTER_PASS_PROPAGATE:                                   \
-            fail += _suiteName##_testSuiteRunner(cond);                        \
+            fail += _PICOTEST_CONCATENATE(_PICOTEST_NAMESPACED(_suiteName),    \
+                                          _testSuiteRunner)(cond);             \
             break;                                                             \
         case PICOTEST_FILTER_SKIP:                                             \
             break;                                                             \
         case PICOTEST_FILTER_SKIP_PROPAGATE: {                                 \
-            PicoTestMetadata **subtest = _suiteName##_subtests;                \
+            PicoTestMetadata **subtest = _PICOTEST_CONCATENATE(                \
+                _PICOTEST_NAMESPACED(_suiteName), _subtests);                  \
             for (; *subtest; subtest++) {                                      \
                 fail += (*subtest)->test(cond);                                \
             }                                                                  \
@@ -1687,7 +1781,11 @@ typedef void(PicoTestSuiteAfterSubtestProc)(const char *suiteName, int nb,
 /**
  * Turn argument into a C string.
  */
-#define _PICOTEST_STRINGIZE(arg) #arg
+#define _PICOTEST_STRINGIZE(arg) _PICOTEST_STRINGIZE2(arg)
+
+/*! \cond IGNORE */
+#define _PICOTEST_STRINGIZE2(arg) #arg
+/*! \endcond */
 
 /**
  * Concatenate both arguments.