+++ docs/html/dir_d44c64559bbebec7f509842c48db8b23.js
@@ -1,4 +1,5 @@
 var dir_d44c64559bbebec7f509842c48db8b23 =
 [
-    [ "picotest.h", "picotest_8h.html", "picotest_8h" ]
+    [ "picotest.h", "picotest_8h.html", "picotest_8h" ],
+    [ "picotestRunner.inc", "picotest_runner_8inc.html", "picotest_runner_8inc" ]
 ];