+++ docs/html/navtreeindex0.js
@@ -112,6 +112,8 @@ var NAVTREEINDEX0 =
 "pages.html":[],
 "picotest_8h.html":[3,0,1,0],
 "picotest_8h_source.html":[3,0,1,0],
+"picotest_runner_8inc.html":[3,0,1,1],
+"picotest_runner_8inc.html#a5f1501e3294d9a3c819fbd0863d2878f":[3,0,1,1,0],
 "tags_8c.html":[3,0,0,5],
 "topics.html":[1],
 "traverse_8c.html":[3,0,0,6]