+++ docs/html/search/files_4.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['picotest_2eh_0',['picotest.h',['../picotest_8h.html',1,'']]]
+  ['picotest_2eh_0',['picotest.h',['../picotest_8h.html',1,'']]],
+  ['picotestrunner_2einc_1',['picotestRunner.inc',['../picotest_runner_8inc.html',1,'']]]
 ];