+++ docs/html/search/searchdata.js
@@ -7,8 +7,9 @@ var indexSectionsWithContent =
   4: "p",
   5: "p",
   6: "p",
-  7: "acfipst",
-  8: "acdefilmptuw"
+  7: "p",
+  8: "acfipst",
+  9: "acdefilmptuw"
 };
 
 var indexSectionNames =
@@ -20,8 +21,9 @@ var indexSectionNames =
   4: "typedefs",
   5: "enums",
   6: "enumvalues",
-  7: "groups",
-  8: "pages"
+  7: "defines",
+  8: "groups",
+  9: "pages"
 };
 
 var indexSectionLabels =
@@ -33,7 +35,8 @@ var indexSectionLabels =
   4: "Typedefs",
   5: "Enumerations",
   6: "Enumerator",
-  7: "Modules",
-  8: "Pages"
+  7: "Macros",
+  8: "Modules",
+  9: "Pages"
 };